move tests near the generator
