+++ generators/spring-service/spring-service.spec.mts
@@ -1,9 +1,7 @@
 import assert from 'yeoman-assert';
-import fse from 'fs-extra';
-import { skipPrettierHelpers as helpers } from './support/index.mjs';
-import { SERVER_MAIN_SRC_DIR } from '../generators/generator-constants.mjs';
-import { GENERATOR_SPRING_SERVICE } from '../generators/generator-list.mjs';
-import { getTemplatePath } from './support/index.mjs';
+import { skipPrettierHelpers as helpers } from '../../test/support/index.mjs';
+import { SERVER_MAIN_SRC_DIR } from '../generator-constants.mjs';
+import { GENERATOR_SPRING_SERVICE } from '../generator-list.mjs';
 
 describe('generator - service', () => {
   describe('creates service without interface', () => {
